move drawable pngs into a drawable-nodpi res directory.
the `drawable` directory is equivalent to drawable-mdpi which is not
where we should place drawable png's that are not meant to scale with
device size. This can quickly lead to crashes where the app tries to
draw an excessively large bitmap.

Instead, we put these drawables in a `nodpi` drawable directory which is
where drawables that are not meant to be scaled should go.

Fixes #378